home page primary

+++ server/sqlSchame/workshop_passage.sql
@@ -1,5 +1,5 @@
 create table workshop_passage (
 	passage_id bigint references passages(id),
-	workshop_id bigint references workshop(id),
+	workshop_id integer references workshop(id),
 	primary key (passage_id)
 )
